refactor: create utils/position package
